추가: US/EDGAR 판독 실배선 (tableUs) + 시장 라우팅 (KR/US wired)
tableUs 가 EDGAR 원천(prices OHLCV·finance XBRL frame·allFilings form)을 KR 대칭 스키마로 직독.
markets.US=wired + tableModule 라우팅, readingCycle·runweek 시장 파라미터화, costs 시장별 틱/세율
(US 1센트 틱 + SEC fee). 실데이터 검증: US runWeek 601주·7683종목 발행, certify 로 fund.bm t5.5
인증(US 가치 팩터)·price.ret5 t-6.4 발굴(US 단기 리버설)·high52/volShock 인증 = US 시장구조 재현
(KR과 달리 가격엣지 존재 = 시장 내 완결 설계 작동). week Int64 캐스트(abstain vstack 가드).

+++ src/dartlab/simulate/table.py
@@ -54,7 +54,7 @@ def weekCalendar(baseDir: Path | None = None) -> tuple[pl.DataFrame, pl.DataFram
         .unique()
         .sort("date")
         .with_columns(d=pl.col("date").str.to_date("%Y%m%d"))
-        .with_columns(week=(pl.col("d").dt.iso_year() * 100 + pl.col("d").dt.week()))
+        .with_columns(week=(pl.col("d").dt.iso_year() * 100 + pl.col("d").dt.week()).cast(pl.Int64))
     )
     weekMap = cal.select("date", "week")
     weekEnd = cal.group_by("week").agg(pl.col("date").max().alias("date"))
